+++ packages/ddp-client/package.js
@@ -15,6 +15,9 @@ Package.onUse(function (api) {
            'retry', 'id-map', 'ecmascript'],
           ['client', 'server']);
 
+  // for DDP.onReconnect
+  api.use('callback-hook', ['client', 'server']);
+
   // common functionality
   api.use('ddp-common', ['client', 'server']);